% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_interpreter.R
\name{aggregateResults}
\alias{aggregateResults}
\title{Aggregation of Single BBoB Results}
\usage{
aggregateResults(allResults)
}
\arguments{
\item{allResults}{The user has to pass a number of result objects produced by applying \code{\link{readOutput}}.}
}
\value{
\code{aggregateResults} returns a list that contains the following components aggregated over all passed single bbob results
(see \code{\link{readOutput}} for information on the non aggregated single bbob results):
  \item{aggregatedAllBest}{vector of all best fitness values.}
  \item{aggregatedAvgBest}{aggregated average best fitness values.}
  \item{aggregatedOverallBest}{overall best fitness values.}
  \item{aggregatedOverallWorst}{overall worst fitness values.}
  \item{aggregatedSDBests}{standard deviations of all best fitness values.}
  \item{aggregatedAllRuns}{vector of the number of iterations.}
  \item{aggregatedLongestRun}{overall highest required number of iterations.}
  \item{aggregatedShortestRun}{overall lowest required number of iterations.}
  \item{aggregatedAvgRun}{average required iterations.}
  \item{aggregatedSDRuns}{standard deviation of required iterations (\code{sd(aggregatedAllRuns})).}
  \item{aggregatedAllRunsEval}{vector of the number of function evaluations.}
  \item{aggregatedLongestRunEval}{overall highest required number of function evaluations.}
  \item{aggregatedShortestRunEval}{overall lowest required number of function evaluations.}
  \item{aggregatedAvgRunEva}{average required function evaluations.}
  \item{aggregatedSDRunsEval}{standard deviation of required function evaluations (\code{sd(aggregatedAllRunsEval})).}
  \item{aggregatedAllStagnation}{vector of the final iterations without improvement.}
  \item{aggregatedLongestStagnation}{overall highest number of iterations without improvement.}
  \item{aggregatedShortestStagnation}{overall number of iterations without improvement.}
  \item{aggregatedAvgStagnation}{average number of final iterations without improvement in the fitness value.}
  \item{aggregatedSDStagnation}{standard deviation of the number of final iterations without improvement 
  (sd of \code{aggregatedAllStagnation}).}
  \item{aggregatedAllConvergence}{a \code{\link{data.frame}} that stores the average convergence of every single bbob result.}
  \item{aggregatedAvgConvergence}{the mean of the convergence of all single bbob results (\code{mean(aggregatedAllConvergence})).}
  \item{aggregatedAllRestarts}{vector of the number of restarts.}
  \item{aggregated_t_test_termination}{vector of the number of t-test terminations while using OCD as a stopping condition.}
  \item{aggregated_chi_test_termination}{vector of the number of chi\code{^2}-test terminations while using OCD as a stopping condition.}
}
\description{
\code{aggregateResults} is used to aggregate the results generated by \code{\link{readOutput}}.
}
\details{
The function \code{aggregateResults} takes a number of result objects, produced by applying
\code{readOutput} on single data files, and aggregates those results. 
For example: The user can pass the results of two function optimization runs. First, he generates the results of each single output by
calling \code{\link{readOutput}} for every single file (e.g. \code{result1 = }\code{\link{readOutput}}(\code{output_function1}), 
\code{result2 = }\code{\link{readOutput}}(\code{output_function2}). 
Then, the user passes these results to \code{\link{aggregateResults}} 
(e.g. \code{\link{aggregateResults}}(\code{c(result1, result2}))
}

