% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_interpreter.R
\name{allresults_processing}
\alias{allresults_processing}
\alias{getAggregatedConvergenceFunctions}
\alias{getAvgBestPerDimension}
\alias{getAvgBestPerFunction}
\alias{getAvgBestPerFunctionAndDimension}
\title{Interpretation of Aggregated BBoB Results}
\usage{
getAggregatedConvergenceFunctions(results, nFunctions, nDimensions)

getAvgBestPerFunctionAndDimension(results, nFunctions, nDimensions)

getAvgBestPerFunction(results, nFunctions, nDimensions)

getAvgBestPerDimension(results, nFunctions, nDimensions)
}
\arguments{
\item{results}{\code{results} must be a return object of \code{\link{aggregatedResults}}, i.e. the aggregated results of several function optimizations.}

\item{nFunctions}{the number of function for which data exists and results should be computed}

\item{nDimensions}{the number of dimensions for which data exists and results should be computed (nDimensions is the total number of logged dimensions,
i.e. \code{nDimensions} has to be a counting value, not the actual dimensionality)}
}
\value{
The return objects are function specific and as follows:
  \item{aggregatedConvergenceFunctions}{A matrix of the fitness gaps for each function averaged over all dimensions.}
  \item{getAvgBestPerFunctionAndDimension}{A vector that stores the average best value (mean over all instances)
  for each function and each dimension of that function.}
  \item{getAvgBestPerFunction}{A vector that stores the average best value (mean over all instances) for each function over
  (dimensions are not considered separately.}
  \item{getAvgBestPerDimension}{A vector that stores the average the best value
  for each dimension over all functions (functions are not considered separately).}
}
\description{
\code{allresults_processing} is a collection of functions for interpreting aggregated bbob results 
(i.e. a return object of \code{\link{aggregateResults}}).
}
\details{
The functions as defined above serve the following purposes:
\describe{
  \item{\code{getAggregatedConvergenceFunctions}}{A function that averages the convergence 
  for each optimized function over all dimensions specified.}
  \item{\code{getAvgBestPerFunctionAndDimension}}{A function that averages the best results
  for each function and all dimensions specified.}
  \item{\code{getAvgBestPerFunction}}{A function that averages the best results
  for each function over all dimensions specified.}
  \item{\code{getAvgBestPerDimension}}{A function that averages the best results
  for each dimension over all functions specified.}
  }
}

