% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_interpreter.R
\name{averageConvergence}
\alias{averageConvergence}
\title{Average Convergence per Function and/or per Dimension}
\usage{
averageConvergence(allConvergence, includedFunctions, includedDimensions,
  nDimensions)
}
\arguments{
\item{allConvergence}{TODO}

\item{includedFunctions}{TODO}

\item{includedDimensions}{TODO}

\item{nDimensions}{the number of dimensions for which data exists and results should be computed (nDimensions is the total number of logged dimensions,
i.e. \code{nDimensions} has to be a counting value, not the actual dimensionality)}
}
\value{
TODO
}
\description{
average convergence per function or per dimension or per a combination of both.
}

