% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customized_bbob.R
\name{bbo_benchmarking_optimizers}
\alias{bbo_benchmarking_optimizers}
\alias{optimizerCMAES}
\alias{optimizerCMAESWithoutDef}
\alias{optimizerGA}
\alias{optimizerRS}
\title{Optimizer for BBo_Benchmarking}
\usage{
optimizerCMAES(dimension, instance, function_id, maxit, maxFE, stopFitness,
  path, debug.logging = FALSE, max_restarts, restart_multiplier,
  restart_triggers, OCD = FALSE, varLimit = NULL, nPreGen = NULL,
  maxGen = NULL, fitnessValue = FALSE, dispersion = FALSE,
  evolutionPath = FALSE)

optimizerCMAESWithoutDef(dimension, instance, function_id, maxit, maxFE,
  stopFitness, path, debug.logging = FALSE, max_restarts, restart_multiplier,
  restart_triggers, OCD = FALSE, varLimit = NULL, nPreGen = NULL,
  maxGen = NULL, fitnessValue = FALSE, dispersion = FALSE,
  evolutionPath = FALSE)

optimizerRS(dimension, instance, function_id, maxit, maxFE, stopFitness, path,
  OCD = FALSE, debug.logging = FALSE, max_restarts = 0,
  restart_multiplier = 1, restart_triggers = character(0), ...)

optimizerGA(dimension, instance, function_id, maxit, maxFE, stopFitness, path,
  debug.logging = FALSE, max_restarts, restart_multiplier, restart_triggers,
  OCD = FALSE, varLimit = NULL, nPreGen = NULL, maxGen = NULL,
  fitnessValue = FALSE, dispersion = FALSE, evolutionPath = FALSE)
}
\arguments{
\item{maxit}{[\code{integer}]\cr
If \code{maxFE} is not passed, the number of iterations, \code{maxit}, serves as an upper execution limit for optimizing one function.}

\item{maxFE}{[\code{interger}]\cr
\code{maxFE} is another upper execution limit for an optimizer, i.e. if the specified number of function evaluations is reached,
the optimizer terminates the optimization.}

\item{stopFitness}{[\code{numeric}]\cr
If \code{stopFitness} is specified, the optimization is terminated if the gap between the current function
value and the global optimum of the function is below the value of \code{stopFitness}.}

\item{debug.logging}{[\code{logical}]\cr
\code{debug.logging} ...
Default is \code{FALSE}.}

\item{OCD}{[\code{logical}]\cr
\code{OCD} indicates if Online Convergence Detection (OCD) should be used as a stopping condition of the
optimizer under test. OCD can only be activated if the optimizer has implemented the functionality necessary for OCD.
See \code{\link{stopOnOCD}} for further necessary parameters.
Default is \code{FALSE}.}

\item{varLimit}{[\code{logical}]\cr
\code{varLimit}: OCD Parameter (See \code{\link{stopOnOCD}})}

\item{nPreGen}{[\code{logical}]\cr
\code{nPreGen}: OCD Parameter (See \code{\link{stopOnOCD}})}

\item{maxGen}{[\code{logical}]\cr
\code{maxGen}: OCD Parameter (See \code{\link{stopOnOCD}})}

\item{fitnessValue}{[\code{logical}]\cr
\code{fitnessValue}: OCD Parameter (See \code{\link{stopOnOCD}})}

\item{dispersion}{[\code{logical}]\cr
\code{disperion}: OCD Parameter (See \code{\link{stopOnOCD}})}

\item{evolutionPath}{[\code{logical}]\cr
\code{evolutionPath}: OCD Parameter (See \code{\link{stopOnOCD}})}

\item{dimensions}{[\code{integer}]\cr
The \code{dimensions} parameter determines the dimensions to be passed to the optimizer under test. Possible values are between
2 and 40.}

\item{instances}{[\code{integer}]\cr
Every \code{smoof_function} can be instantiated in different ways. Could be a single intergerish value 
greater of equals 1.}

\item{function_ids}{[\code{integer(1)}]\cr
The \code{function_ids} are the unique identifier of the 24 noiseless bbob function.}

\item{max.restarts}{[\code{integer}]\cr
If \code{max.restarts} the optimizer restarts the optimization if one of the passed \code{restart_triggers} has fired.
In order to use this feature, the optimizer under test must be able to restart the optimization.
Default is \code{FALSE}.}

\item{restart.triggers}{[\code{string}]\cr
If \code{restart.triggers} are passed, the optimizer restarts the optimization if one of those triggers has fired.
In order to use this feature, the optimizer under test must be able to restart the optimization.
Default is \code{FALSE}.}
}
\description{
Optimizers ready for usage in bbob_custom or bbob_custom parallel
}
\details{
The optimizers are ready to be set up in a benchmarking experiment with \code{\link{bbob_custom}} or
\code{\link{bbob_custom_parallel}}. Every input parameter is required as it passed by
\code{\link{bbob_custom}} or \code{\link{bbob_custom_parallel}} to the particular optimizer.
}

