% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis_function.R
\name{createOutputImages}
\alias{createOutputImages}
\title{Output Analysis of CMAES}
\usage{
createOutputImages(outputPath)
}
\arguments{
\item{outputPath}{The user has to define the desired path \code{outputPath}. A folder will be created and filled with analysis results.}
}
\value{
\code{outputPath} creates a folder on harddrive that contains every image or result of the benchmarking experiments conducted to 
compare the default stopping criterion of CMAES with OCD as a criterion.
}
\description{
\code{createOutputImages} creates all relevant analysis results in form of images.
}
\details{
\code{createOutputImages} will only work, if every necessary file is present in the working directory.
}

