% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_interpreter.R
\name{extractECDFofFunctions}
\alias{extractECDFofFunctions}
\title{Empirical Cumulative Distribution Function of BBoB data}
\usage{
extractECDFofFunctions(results, fitnessGap = 1e-08)
}
\arguments{
\item{results}{\code{results} must be a return object of \code{\link{aggregatedResults}}, i.e. the aggregated results of several function optimizations.}
}
\value{
\code{extractECDFofFunctions} returns an empirical cumulative ditribution function based on bbob data which can be plotted directly.
}
\description{
\code{extractECDFofFunctions} returns a cumulative distribution function of 
the functions that were solved within the desired fitness gap by the number of function evaluations this took
}
\details{
\code{extractECDFofFunctions} can be used to observe which fraction of functions has been solved within the desired fitness gap
passed to the function. For example, a bbob experiment could terminate an optimization run after 100000 function evaluations.
If the specified fitness gap is not reached after this number of function evaluations, the corresponding function remains unsolved
with respect to that specific gap.
}

