% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_interpreter.R
\name{getActiveFunctions}
\alias{getActiveFunctions}
\title{Get the Number of Active Functions in an Interation}
\usage{
getActiveFunctions(results)
}
\arguments{
\item{results}{\code{results} must be a return object of \code{\link{aggregatedResults}}, i.e. the aggregated results of several function optimizations.}
}
\value{
\code{getActiveFunctions} returns a vector that contains the number of functions (the number of instances of all functions) 
that are not stopped (per iteration).
}
\description{
\code{getActiveFunctions} returns the number of functions per iteration that are not yet stopped.
Functions stop e.g. when a certain solution quality is reached
}

