% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMonitor.R
\name{makeTXTMonitor}
\alias{makeTXTMonitor}
\title{Generator for .txt output monitor.}
\usage{
makeTXTMonitor(max.params = 4L, path, Fopt, function_id, dimension,
  instance_id)
}
\arguments{
\item{max.params}{[\code{integer(1)}]\cr
Maximal number of parameters to show in output.}

\item{path}{Path to output}

\item{Fopt}{Global optimum of function}
}
\value{
[\code{cma_monitor}]
}
\description{
This monitor generates a txt file used for result logging
}

