% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbga.R
\name{rbga}
\alias{rbga}
\title{export}
\usage{
rbga(stringMin = c(), stringMax = c(), suggestions = NULL,
  popSize = 200, iters = 100, mutationChance = NA, elitism = NA,
  monitorFunc = NULL, evalFunc = NULL, showSettings = FALSE,
  verbose = FALSE, instance, control = list())
}
\description{
export
}

