% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_interpreter.R
\name{readOutput}
\alias{readOutput}
\title{Interpretation of BBoB Data}
\usage{
readOutput(file)
}
\arguments{
\item{file}{The user has to specify the exact datapath to the bbob data to be read and interpreted.}
}
\value{
\code{readOutput} returns an object of the class \code{single_bbob_result}. A list of those can be passed to
\code{\link{aggregateResults}}. Each \code{single_bbob_result} is a list that contains the following components:
  \item{allBest}{best fitness value of each processed function instance.}
  \item{avgBest}{average best fitness value of all processed function instances.}
  \item{overallBest}{overall best fitness value of all processed function instances.}
  \item{overallWorst}{overall worst fitness of all processed function instances.}
  \item{sdbest}{standard deviation of all best fitness values (\code{sd(allBest)}).}
  \item{allRuns}{number of iterations of each processed function instance that has been required for optimization.}
  \item{longestRun}{overall highest required number of iterations of all function instances.}
  \item{shortestRun}{overall lowest required number of iterations of all function instances.}
  \item{avgRuns}{average required iterations of all functions instances.}
  \item{sdRuns}{standard deviation of required iterations of all function instances (\code{sd(allRuns})).}
  \item{allRunsEval }{number of function evaluations of each processed function instance that has been required for optimization.}
  \item{longestRunEval}{overall highest required number of function evaluations of all function instances.}
  \item{shortestRunEval}{overall lowest required number of function evaluations of all function instances.}
  \item{avgRunEval}{average required function evaluations of all functions instances.}
  \item{sdRunsEval}{standard deviation of required function evaluations of all function instances (\code{sd(allRunsEval})).}
  \item{allStagnations}{number of final iterations without improvement in the fitness value.}
  \item{longestStagnation}{highest number of iterations without improvement in the fitness value of all function instances.}
  \item{shortestStagnation}{lowest number of iterations without improvement in the fitness value of all function instances..}
  \item{avgStagnation}{average number of final iterations without improvement in the fitness value.}
  \item{sdStagnation}{standard deviation of the number of final iterations 
  without improvement in the fitness value of all function instances (\code{sd(allStagnation})).}
  \item{allConvergence}{a \code{\link{data.frame}} that stores the gap between the fitness value and the global optimal value for each
  iteration.}
  \item{avgConvergence}{a vector that stores \code{mean(allConvergence)} of all processed instances applied 
  over all iterations.}
  \item{allRestarts}{number of restarts occured in optimizing one single function instance.}
  \item{t_test_termination}{number of t-test terminations while using OCD as a stopping condition.}
  \item{chi_test_termination}{number of chi\code{^2}-test terminations while using OCD as a stopping condition.}
  \item{functionID}{the identifier of the bbob function that has been optimized.}
  \item{dimension}{the problem dimensions for the optimization.}
}
\description{
\code{readOutput} is used to read and interpret an output file generated by the benchmarking functions
\code{\link{bbob_custom}} or \code{\link{bbob_custom_parallel}}.
}
\details{
There are several functions for reading and interpreting the output generated by \code{\link{bbob_custom}} 
or \code{\link{bbob_custom_parallel}}. \code{readOutput} is used to read one single output file and returns a list of
results of that data (see value section). The user might execute something as follows: 
\code{\link{readOutput}}(\code{output_function1}).
}

