% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnOCD}
\alias{stopOnOCD}
\title{Stopping condition: Online Convergence Detection.}
\usage{
stopOnOCD(varLimit, nPreGen, maxGen = NULL, fitnessValue = TRUE,
  dispersion = FALSE, evolutionPath = FALSE)
}
\arguments{
\item{max.iter}{[integer(1)]\cr
Maximal number of iterations.
Default is \code{100}.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
StopOnOCD.
}
\seealso{
Other stopping.conditions: \code{\link{stopOnCondCov}},
  \code{\link{stopOnIndefCovMat}},
  \code{\link{stopOnMaxIters}},
  \code{\link{stopOnNoEffectAxis}},
  \code{\link{stopOnNoEffectCoord}},
  \code{\link{stopOnOptParam}},
  \code{\link{stopOnOptValue}},
  \code{\link{stopOnTimeBudget}}
}

