% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_interpreter.R
\name{averageConvergence}
\alias{averageConvergence}
\title{Average Convergence per Function and/or per Dimension}
\usage{
averageConvergence(allConvergence, includedFunctions, includedDimensions,
  nDimensions)
}
\arguments{
\item{allConvergence}{matrix of convergence values}

\item{includedFunctions}{functions for which the average convergence should be computed}

\item{includedDimensions}{dimensions for which the average convergence should be computed}

\item{nDimensions}{the number of dimensions for which data exists and results should be computed (nDimensions is the total number of logged dimensions,
i.e. \code{nDimensions} has to be a counting value, not the actual dimensionality)}
}
\value{
\code{averageConvergence} returns the average of the convergence values for the functions and/or dimensions specified when calling the function.
}
\description{
\code{averageConvergence} returns a matrix with the convergence values per function or per dimension or per a combination of both.
}

